VOIP-522: USB OTG DWC2 driver: The parity fix in isochronous mode
USB OTG driver in isochronous mode has to set the parity of the receiving
microframe. The parity is set to even by default. This causes problems
for an audio gadget, if the host starts transmitting on odd microframes.

This fix uses Incomplete Periodic Transfer interrupt to toggle between
even and odd parity until the Transfer Complete interrupt is received.

Bug: VOIP-522
Change-Id: Icbd11e58f4acaa8f20c1e8337a392fcd9964c163
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/8716
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Roman Bacik <rbacik@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>