LINUXDEV-139: Use clock instead of specifying UART speed in NSP DT
ARM: dts: NSP: Use clock instead of specifying UART speed

Use the osc clock instead of hard coding the speed for the UART.  The
speed being set is the same, but this is "more correct".

NOTE: This is a modification to the pre-existing behavior, and requires
an updated u-boot. (DO NOT UPSTREAM THIS COMMENT).

Signed-off-by: Jon Mason <jonmason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: I2f5c968a5651b7c177c97954132a34abc72ef1b6
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/11236
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>