VOIP-5085 net: ethernet: Broadcom AMAC Ethernet Driver for IPROC SoC
This patch adds the AMAC Ethernet driver support for the
internal Ethernet Controller in the Broadcom IPROC family
SoC's.

This patch introduces support for the following Broadcom
Cygnus SoCs:
 - BCM583XX (BCM58300, BCM58302, BCM58303, BCM58305)
 - BCM113XX (BCM11300, BCM11320, BCM11350, BCM11360)

The driver is implemented without any switch support. Switch
support will be added in a later commit. The driver currently
runs in the switch-by-pass mode. In this mode the MAC can
directly control Port0. So only Port 0 is supported with
this patch.

Bug: VOIP-5085
Change-Id: I7f3d121ef9ed03f42fd5cc7f4a1243680d88172e
Signed-off-by: Arun Parameswaran <arunp@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/11671
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>