LINUXDEV-139: arm: dts: Add syscon based reboot in NSP DT
arm: dts: nsp: Add syscon based reboot in DT

Add the ability to reboot via a reset of the processor.  This is
achieved via a write of 0x2f9 to the CRU Reset Register.  Unfortunately,
this only resets the core and not the other IP blocks.  So if possible,
other methods should be used on the individual boards.

Signed-off-by: Jon Mason <jonmason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: I5e5c57949ba7e36cccba8d9665ea69bacda267d9
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/13593
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>