CTRL-20951: Use node labels to override attributes in NS2 SVK DT
arm64: dts: Use node labels to override attributes in NS2 SVK DT

It is recommended to use node label when overriding contents of
a node in .dts file. Doing so will avoid accidental creation of
new DT nodes in future when the actual DT node is renamed in the
included .dtsi file.

This patch updates ns2-svk.dts as-per above recommendation.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: CTRL-20951
Change-Id: I20e74b92bf8ac129d3754dab4a6c0ff6069c0c3f
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/13900
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>
Reviewed-by: Vikram Prakash <vikramp@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>