LINUXDEV-139: Correct sp804 timer address
ARM: dts: NSP: Correct sp804 timer address

It appears that the sp804 address was not properly updated when the
range translation was added to this.  This is resulting in a WARN_ON
being triggered duing the boot of the kernel.  Correction of the address
to the correct value results in the error going away.  This isn't an
issue upstream, as this device tree entry has not been pushed yet.

Signed-off-by: Jon Mason <jonmason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: I1033998f957c6f2d629d9d8da8aa7986283a4d3c
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/16004
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>