LINUXDEV-139: Add PCI_DOMAIN dependency to PCI Kconfig
PCI: iproc: Add PCI_DOMAIN dependency to PCI Kconfig

2+ PCI devices fail to be discovered due to each bus having the same
PCI domain.  This is because the domain defined in the device tree file
is not being added due to PCI_DOMAIN not being enabled.  So, every PCI
bus has a domain of zero.  When PCI_DOMAIN is selected by the Kconfig,
it picks-up the domain defined in the device tree file and everything
works as expected.

Signed-off-by: Jon Mason <jonmason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: Icdd2a14d0bef84cf11f4479b7e351ec04183fe54
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/16033
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>