LINUXDEV-139: NSP MSI/INTx IRQ collision
ARM: dts: NSP: MSI/INTx IRQ collision

It is possible for a PCI device to use the PCI bus' MSI and INTx
interrupts simultaniously (via a PCI bridged device).  Given this
possibility, the INTx interrupt should be removed from the list of
potential interrupts in the NSP device tree.  Also, due to the
sequential way the driver allocates the interrupts, the vector
following the INTx vector must also be removed.

Signed-off-by: Jon Mason <jonmason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: I6eea45136d4f68a7c2b3d4b4ed6d9ac1394c8747
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/16090
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <rjui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>