CTRL-20886: Use SCLK as input clock for UART on Stingray
arm64: dts: Use SCLK as input clock for UART on Stingray

On Stingray, we can use PCLK (87.5 MHz) or SCLK (25 MHz)
as input clock for UARTs. By default upon reset, the UART
input clock is set to PCLK. Other iProc SoCs such as NS2
and Pegasus use SCLK as input clock for UARTs.

This patch updates "clock-frequency" property of UART DT
node assuming SCLK as input clock. It will be responsiblity
of arm-trusted-firmware to ensure SCLK is selected as input
clock for all UARTs on Stingray.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: CTRL-20886
Change-Id: Ifb12783f1fd8c87f421ee75ce92fdabe995ff733
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/17540
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>