VOIP-5085 net: ethernet: brcm-amac: Convert switch & IO pad reg to optional params
Converted Switch and IO Pad registers in AMAC driver to
optional params. These are required to be specified only
for certain SoC's.

Modified Switch and IO Pad configurations to be based on the
availability of these regs.

Modified the switch mode variable to be dependent on the
availability of 'brcm,enet-switch-mode' in DT to indicate the
availability of an internal switch.

When the configuration is absent in DT, the AMAC works in
switch-by-pass mode.

Bug: VOIP-5085
Change-Id: Ic1598077810ab7b948b182e8f9a41d7b7e51e3e4
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/18178
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Tested-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>