SOC-3278:  Make use of SPU hw register
crypto: brcm - Update Broadcom SPU driver

The device tree entry for the SPU driver includes the SPU register
ranges. Previously, the driver read the ranges to determine the
number of SPU hw blocks, but did not remap the ranges or read or
write any of the registers. The driver now reads a fifo high
water mark and prints it with other stats in the debug fs output.

Bug: SOC-3278
Change-Id: I304e7970cb05f7f0d9bdfa9250877e059fdc89bb
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/21772
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Pramod KUMAR <pramod.kumar@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>