CTRL-22067: Add Rabin Finger Print algorithm
crypto:rabin: Add Rabin Finger Print algorithm

This algorithm uses content-defined chunking with the help of rolling hash,
a hash function where the input is hashed in a window that moves through
the input.

it divides the input stream into chunks based on rolling hash and calculate
the hash of every chunk. it produces the list of descriptors having-
1. Hash of the created chunk
2. Length of this chunk.
3. Offset of the chunk in input stream.
4. Fingerprint of this chunk.
5. status indicating validity of this descriptor

Bug: CTRL-22067
Change-Id: I2d46a126160240c58621f1f22f1d745e3e931c1a
Signed-off-by: Pramod Kumar <pramod.kumar@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/22237
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>