CTRL-22268: pcie fixup for classless AVIP
arm64: pcie fixup for classless AVIP

for palladium, we have AVIP connected to core1.
the end point (AVIP) does not get enuramated because
it is class-less device.

linux pci framework doesi not align the mem and io resources
for the classless device.

we force the class to be of memory class for AVIP for
both fn1 and fn2.

DO NOT UPSTREAM !!!

Bug: CTRL-22268
Change-Id: I5240a63e691c458b5363cd5e43533d619c0f0d3d
Signed-off-by: Oza Pawandeep <oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/24058
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>