CTRL-20886:Add header file for SHA3 algorithm's
Currently linux kernel doesn't have software implementation
of SHA3 algorithm. There is one RFC implementing SHA3 but
it's not merged to kernel yet. SPU2 engine in stingray supports
SHA3 crypto offload functionality. To add support for SHA3 in
SPU2 crypto driver we needed to add generic defines in a header
file and include it in the driver. So adding this header file.

Bug: CTRL-20886
Change-Id: Ieb34caaa65222163e170fa8c7664e486abb06922
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/24112
Tested-by: RAVEENDRA PADASALAGI <raveendra.padasalagi@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>