CTRL-20886: Few optimizations in message to descriptors conversion
mailbox: flexdma: Few optimizations in message to descriptors conversion

This patch does following optimizations in message to descriptors
conversion:
1. The constraint of having atleast 2 non-header descriptors after
a header descriptor is not required. The HW has been fixed for this.
2. Pre-compute number of non-header descriptors and pass that as
argument to flexdma_write_descs() so that we don't have to re-compute
it at time of writing descriptors.
3. Merge flexdma_check_n_enqueue_header() funciton into
flexdma_enqueue_desc() function to simplify flexdma_spu_write_descs()
and flexdma_sba_write_descs()

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: CTRL-20886
Change-Id: I00e5373848620dfcc3e58efd9933f32694322429
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/24192
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>