CTRL-20886: Fix max-pq-disks DT attribute for Stingray A0
arm64: dts: Fix max-pq-disks DT attribute for Stingray A0

We will not be using packet extension for FS4 RAID engine
on Stingray A0 because FS4 packet extension cannot be used
along with FS4 HW scheduler.

This means we have to pack all types of RAID requests in
31 descriptors and we can only support max 14 XOR sources
and max 12 PQ sources in-absence of packet extension on
Stingray.

We hope that this limitation in using FS4 packet extension
along with FS4 HW scheduler will be removed in Stingray B0.

This patch fixes max-pq-disks DT attribute as-per above in
Stingray DT.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: CTRL-20886
Change-Id: Ia4810586a591f783e1094804c800062a77fec13a
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/24648
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>