SOC-3555: Broadcom USB DRD Phy driver for Northstar2
This is driver for USB DRD Phy used in Broadcom's Northstar2
SoC. The phy can be configured to be in Device mode or Host
mode based on the type of cable connected to the port. The
driver registers to  extcon framework to get appropriate
connect events for Host/Device cables connect/disconnect
states based on VBUS and ID interrupts.

Bug: SOC-3555
Change-Id: I431ea5930d7f3f6c9d7ccfcf38cbef51788d21b0
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/24739
Reviewed-by: Raviteja Garimella <raviteja.garimella@broadcom.com>
Tested-by: Raviteja Garimella <raviteja.garimella@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Yendapally Reddy Dhananjaya Reddy <yendapally.reddy@broadcom.com>
Reviewed-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>