SOC-5605: phy: phy-bcm-ns2-usbdrd: Broadcom USB DRD PHY driver for Northstar2
Initial version of DRD PHY driver that's accepeted by upstream
maintainer.

This is driver for USB DRD PHY used in Broadcom's Northstar2
SoC. The phy can be configured to be in Device mode or Host
mode based on the type of cable connected to the port. The
driver registers to  extcon framework to get appropriate
connect events for Host/Device cables connect/disconnect
states based on VBUS and ID interrupts.

Change-Id: I2c73e7296c6676e949cbe828129d61ce0ac735cb
Signed-off-by: Raviteja Garimella <raviteja.garimella@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/44891
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>