CTRL-20886: Add hardware HMAC in crypto driver.
Currently crypto driver supports implementation of HMAC
algorithm in software as it's not verified yet on SPUM
with complete implementation in hardware. SPU2 IP has
support for full HMAC algorithm sequence implemented in
hardware. So this patch enhances crypto driver to make
use of SPU2 IP capability for HMAC and keeps the sofware
implementation for SPUM.

Change-Id: I725744cd253744dda29951ff6fa65b8311cea8eb
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Bug: CTRL-20886
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/25812
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>