SOC-3598: In SPU driver, ensure mailbox sgs are deallocated
crypto: brcm - For each request to the SPU hardware, the SPU driver
allocates a mailbox message with two scatterlists. When chaining a
large request, the SPU driver has to dealloc these scatterlists
between each chained request to the hw. Also, only invoke the crypto
API's async callback if the request was handled asynchronously
(i.e., sent to the SPU hw).

Bug: SOC-3598
Change-Id: I52e14d8168f49eba2340f02dc156a7c3fb62119c
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/25642
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: RAVEENDRA PADASALAGI <raveendra.padasalagi@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>