LINUXDEV-139: Move NSP reboot entry to DTS file
ARM: dts: NSP: Move reboot entry to DTS file

The GPIO reboot method is preferred to the core reset method, and all of
the DTS files will be moving to use that.  However, the bcm958625k board
does not have the ability to use GPIO to reset the board.  So, move the
core reset to that DTS file.

Signed-off-by: Jon Mason <jon.mason@broadcom.com>

Bug: LINUXDEV-139
Change-Id: I263e1f5b53194d745f30cce0f74a7189ce1a3ee2
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/25918
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>