SOC-3992: dte support for iproc based SOCs
DTE: iproc: dte support for iproc based SOCs

This driver implements the Broadcom Digital
Timing Engine Driver (DTE). The DTE allows for
hardware time stamping of ivarious clients in a
common adjustable time base.

The DTE creates timestamps of hardware based events such
as GPIO, I2S signals for audio, etc. It provides a clock
for 802.1AS / PTP timestamps for network packets. It has
up to 31 "clients" -- the hardware inputs into the
timestamping engine.

The DTE timestamper creates timestamps based on the
current DTE NCO. When an event on an input occurs,
DTE stores the timestamp in a h/w FIFO. Each client
has a divider that can be set to control the rate that
timestamps are generated at by the timestamper which
are adjustable at run time.

Bug: SOC-3992
Change-Id: Idb116cc5e585d64bdd6e79f5b2b77c4853439588
Signed-off-by: Oza Pawandeep <oza@broadcom.com>
Signed-off-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Signed-off-by: Darren Edamura <darren.edamura@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/28309
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>