CTRL-22268: Inbound memory fix for legacy platforms
PCI: iproc: Inboun memory fix for legacy platforms

this patch fixes the case where legacy platoforms are
not supposed to have dma-ranges.
and if any new platform is added, they are supposed
to have ib_map defined.
in that case we return -EINVAL in order to avoid
dma-range parsing, and as a result of that no inbound memory
will be configured.

Bug: CTRL-22268
Change-Id: Ibf9767a9aa1c9ded45476ea284a551dd304312a1
Signed-off-by: Oza Pawandeep <oza@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/26974
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>