SOC-3881: In SPU driver, fix enforcement of maximum payload
The SPU driver inherited logic to enforce maximum SPU message payload
from an earlier SPU driver. Because of differences in how the two
drivers access hardware, the original logic does not translate directly
to this driver. There was ambiguity about whether there was a limit
to the maximum packet size (SPU message headers plus data) or just
the SPU message payload (just data). Also, the newer SPU2 hardware
does not impose a maximum payload at all, but allows SPU messages of
arbitrary length, learning the total length from the DMA descriptors.

So rework this logic as follows:
- Remove max-pkt-size from the device tree. Remove the validation code
that verified the max-pkt-size was valid.
- Remove the check on total message length when creating mailbox
scatterlists that form a SPU message.
- Remove adjustment of max_payload for AES key size
- Define a special value to indicate infinite max payload
- Remove max_payload from the iproc_alg_s structure and no longer set
max_payload in the driver_algs[] array.
- Define a SPU-hw specific max_payload function to determine the max
payload for a new context. This generally returns INFINITE for SPU2 and
64k (rounded down to a multiple of block size) for SPU-M.

Bug: SOC-3598
Change-Id: Ie6cb9efef5169b3e33e73f9aa617cf0c03772ebe
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/27227
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>