SOC-3598: In SPU driver fix AES CTR chaining
crypto: brcm - When a request is too big to fit in a single SPU
message, the SPU driver has split the request into pieces and
submit a sequence of requests. For AES CTR, the SPU driver has to
update the counter value for each chunk. The counter should be
incremented by the number of AES blocks in the previous chunk,
not the number of bytes.

Bug: SOC-3598
Change-Id: I28f38a6e72059264a71694c743ea145e7182e24d
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/27275
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>