CTRL-22717:arm64-dts:Add QSPI device for stingray
This patch adds a dt node for QSPI controller and NOR flash slave.
Compatible string "brcm,spi-nsp-qspi" used here looks confusing
as it's meant for nsp but the QSPI driver sent to open source is
coded that way. Expecting this to change when the final changes
comes for QSPI driver.

BUG:CTRL-22717
Change-Id: I6694b3a548eae18a349b11c151051bee7febb9c1
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/27533
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>