SOC-3972: SPU driver can't be removed
crypto: brcm - This fix tracks which specific crypto algorithms have
been registered with the crypto API, and when the module is removed,
only those algorithms are unregistered.  Previously, if some algorithms
were not registered (because they did not apply to this SOC type, for
example), they would still be passed to the crypto API unregistration
function at module removal time, leading to a seg fault.

Bug: SOC-3972
Change-Id: I4c39acdfcd9fa1621e08725c1c114c9be22185c3
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/28159
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>