SOC-3414 net: amac: Support for reading mac addr from DT
Some SoC's require support for 'local-mac-address' variable from
u-boot to pass the mac address to the Kernel. They access the
DT variable from u-boot and update it with the correct mac
address.

This patch adds support to read the mac address from the device
tree while maintaining the existing command line argument.
In the absence of a valid mac address via cmd line or DT the
driver will fail to load.

The 'local-mac-address' should be set to '[ 00 00 00 00 00 00 ]'
in the DT. This is an optional property can be modified at
runtime to specify the correct mac address.

Bug: SOC-3414
Change-Id: Ia1ee7fcdde24078c02639da9e2a3c1155461bb29
Signed-off-by: Arun Parameswaran <arun.parameswaran@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/28612
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>