CTRL-20881: Refine code that disables APB error forwarding
PCI: iproc: Refine code that disables APB error forwarding

During enumeration with multi-function EP devices, access to the
configuration space of a non-exist function results in an unsupported
requests being returned as expected. By default the PAXB based iProc
PCIe controller forwards this as an APB error to the host system and
that causes an exception

This issue was originally addressed in:
f0cf2339f24c "PCI: iproc: Fix exception with multi-function EP devices"

It disables this undesired behavior and lets the kernel PCI stack deals
with an access to the non-exist function, in which case vendor ID of
0xffff is returned and handled gracefully

Now there's concern that the disabling of the APB error forwarding
prevents other PAXB APB fatal errors from causing an exception in the
system, e.g., when the host processor tries to access an invalid PAXB
APB register. Therefore, it's more desired that this APB error
forwarding is only disabled during access of the configuration space
registers

Signed-off-by: Ray Jui <ray.jui@broadcom.com>

Bug: CTRL-20881
Change-Id: I0b7fa39da288f16a9f7a91251abed6c8be363653
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/29208
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Oza Oza <oza.oza@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>