SOC-4391: Add AES-CCM support to SPU driver.
This set of changes adds AES-CCM support to the SPU driver, and changes
a few function/variable names to reflect their new usage with CCM. Note
that digest sizes of other than 8, 12, or 16 bytes are not supported by
the SPU hardware, so are handled in software only.

This patch has been cherry-picked from "int_ldk_5.x" branch.

Bug: SOC-4391
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Signed-off-by: Steve Lin <steven.lin1@broadcom.com>

Conflicts:
	Documentation/crypto/bcm-iproc-crypto-readme.txt
	drivers/crypto/bcm/cipher.c
	drivers/crypto/bcm/spu.h
	drivers/crypto/bcm/spu2.h

Change-Id: I972e5f3390ae5ff89bd860b9aa026d710857eae6
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/34984
Tested-by: RAVEENDRA PADASALAGI <raveendra.padasalagi@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>