CTRL-23227: crypto: fix test manager compilation
From 4.6 kernel onwords the definition and reference of
alg_test_pcomp() in crypto/testmanager.c is removed.But
Somehow this piece of code is left out in bcm/master kernel.

When CRYPTO_MANAGER_DISABLE_TESTS is set to N, crypto test
manager includes all the test code and this piece of code is
also included in compilation. Since some of the functions referred
inside alg_test_pcomp() is not defined, compilation error is seen.

By removing the definition of alg_test_pcomp() compilation error
is fixed and which is as per latest kernel code (4.6 onwords).

BUG:CTRL-23227
Change-Id: I1e6dc447882b83da879bd127650043b4b436e758
Signed-off-by: Raveendra Padasalagi <raveendra.padasalagi@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/30144
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Vikram Prakash <vikram.prakash@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>