VOIP-6608 pinctrl: iproc: Add gpio wake source support
A mailbox channel to communicate with the M0 processor is required to
configure a gpio as a wakeup source. The irq_set_wake handler has been
added to send a message to M0 to enable or disable a gpio as a wake
source from sleep or deep sleep.

The irq_disable handler has also been added. The default handler uses a
lazy disable approach where the irq is not masked until it triggers for
the first time. As a result, GPIO interrupts are not physically masked
during PM suspend sequence which can cause a WFI fall-through.

Bug: VOIP-6608
Change-Id: I74d098999fab44136062b6262e1ba5f7dfc3ce62
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/31982
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Tested-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>