SOC-4573: Support for SATA PRBS in Gen1/2/3
One more sysfs entry is created to run the SATA PRBS test in gen1/2/3.
Registers depending on SATA GEN are updated according to the GEN in
which PRBS test is done. Also, some registers updated to make PRBS work
in GEN3. However, for SATA port-2 and 3 (mapped @67d22100 and 67d32100
respectively), GEN3 PRBS has been observed to be failing sometimes.

Bug: SOC-4573
Change-Id: I6db94b2518b63aa96ef58fdff05b562e1817aa5b
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/33105
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Abhishek Rajeshkumar Shah <ABHISHEK.SHAH@BROADCOM.COM>
Reviewed-by: Vikram Prakash <vikram.prakash@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/33329
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>