SOC-4656: Fix interrupt type for generic timer PPIs in Stingray DT
arm64: dts: Fix interrupt type for generic timer PPIs in Stingray DT

The generic timer PPIs are always level-trigered. In fact, Linux-4.9
checks interrupt type of generic timer PPIs and it prints warnings
if it is not level-triggered.

This patch corrects the interrupt type of generic timer PPIs in
Stingray DT.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: SOC-4656
Change-Id: If2b2c779ef5bc09a11493f55befafda18a8bad89
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/34619
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>