SOC-4656: Need console arg in bootargs for Stingray DTs
arm64: dt: Need console arg in bootargs for Stingray DTs

The Linux tty subsystem switches to default console whenever
it is available. If default console is not available then
Linux tty will continue using "bootconsole" until it is
available.

When "console" arg is not specified in bootargs, the default
console is "tty0" whereas it is "ttyS0" when "console=ttyS0"
is specified in bootargs.

The above implies that when "console" arg is not specified
Linux tty subsystem will switch console from "bootconsole"
to "tty0" quite early. Later towards end of booting it will
switch to final console based on linux,stdout-path property
of /chosen DT node. If Linux crashes before we switch to
final console based on linux,stdout-path DT property then
we will not get any prints because console is set to "tty0".
This causes difficulties in debugging.

To make debugging easier for time-being, we should explicity
specify "console" arg in bootargs.

Once we have moved to final bootloader (UEFI) on Stingray,
we can remove "console" arg from bootargs in DTS and instead
use UEFI to specify "console" arg in bootargs.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>
Bug: SOC-4656
Change-Id: Ibcbf222e8c67cde0924ea9b66a060762e7fe5b6a
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/34618
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>