VOIP-6635: Enable the L2C in the Cygnus platform init
arm: mach-bcm:This change configures the secure L2C
Aux control register using the standard ARM smccc call.

This change requires the presence of OPTEE-OS to be running in
the secure world. The OPTEE-OS handler will be trigged by the
smccc call and the handler will call the L2C enable function.

This change combined with the OPTEE-OS in secure world,
will allow the Kernel to boot only in non-secure mode.

Tested-by: ismail <ismail@broadcom.com>
Signed-off-by: ismail <ismail@broadcom.com>

Bug: VOIP-6635
Change-Id: Ic22653b97af15de306f612cec1ca471ff41f34d6
Reviewed-on: http://gerrit-ccxsw.broadcom.net/34799
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>