SOC-4929: BCM54810 PHY support on Stingray
This patch adds Ethernet support for Stingray boards with BCM54810 PHY.
On SR boards, lane-swap is required to get the link up; BCM54810 PHY has
lane-swap support. So this patch does the lane-swap in BCM54810 and makes
sure that all the settings related to BCM54810 are done priorly so tha
data transfer works.

Bug: SOC-4929
Improvement: SOC-4929
Change-Id: I61c64ca399f8f48d50e843724db831afb3e7dddf
Signed-off-by: Abhishek Shah <abhishek.shah@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/34894
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>