VOIP-6477: Decouple interrupt handling from iproc mailbox driver
mailbox: Decouple interrupt handling from iproc mailbox driver

Removes the mailbox interrupt handling from the mailbox driver and adds
it to an interrupt controller driver. These drivers will be upstreamed
as separate patchsets but are included in one commit so that mailbox
support in the CRMU GPIO driver isn't broken.

Bug: VOIP-6477
Change-Id: I412bf1b72d1841fd7778f0dfb56f9ec474bade28
Reviewed-on: http://gerrit-ccxsw.rtp.broadcom.com:8080/35460
Reviewed-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Tested-by: Jonathan Richardson <jonathan.richardson@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: vpx_smoketest status <vpx_smoketest@broadcom.com>
Reviewed-by: CCXSW <ccxswbuild@broadcom.com>
Tested-by: CCXSW <ccxswbuild@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>